refactor(mpi): agrega directiva de foto al módulo
