+++ src/app/app.module.ts
@@ -519,7 +519,6 @@ import { LogoSvgComponent } from './styles/logo.svg';
 import { AcronimoSvgComponent } from './styles/acronimo.svg';
 import { ListaReglasComponent } from './components/top/reglas/listaReglas.component';
 import { VistaSolicitudTopComponent } from './modules/rup/components/huds/vistaSolicitudTop';
-import { FotoDirective } from './components/paciente/paciente-detalle-foto.directive';
 
 import { MPILibModule } from './modules/mpi/mpi-lib.module';
 
@@ -654,7 +653,6 @@ registerLocaleData(localeEs, 'es');
         BusquedaMpiComponent,
         PacienteCruComponent,
         GeorrefMapComponent,
-        FotoDirective,
 
         // form Terapeutico
         FormTerapeuticoComponent,
