+++ src/app/modules/mpi/components/paciente-detalle.html
@@ -7,7 +7,7 @@
     </plex-badge>
 
     <div title>
-        {{paciente | paciente }}
+        {{paciente | nombre }}
     </div>
 
     <div subtitle>
