+++ src/app/modules/mpi/components/paciente-detalle.scss
